feat: initial upgrade to cypress 10, add compatibility for angular-apollo-tailwind
