+++ starter-dev-e2e/cypress/fixtures/user/graphql/currentUser.graphql.json
