+++ starter-dev-e2e/cypress/fixtures/user/graphql/userGist.graphql.json
