+++ starter-dev-e2e/cypress/fixtures/user/graphql/userProfile.graphql.json
