+++ starter-dev-e2e/cypress/fixtures/user/rest/currentUser.json
