+++ starter-dev-e2e/package.json
@@ -1,12 +1,12 @@
 {
   "scripts": {
     "cy:run": "cypress run --config-file $CYPRESS_CONFIG",
-    "start": "START_COMMAND=`node -p \"require('$CYPRESS_CONFIG').startCommand\"` && cd ../$SHOWCASE_APP && $START_COMMAND",
-    "test": "CYPRESS_CONFIG=./cypress/configs/$SHOWCASE_APP.json BASE_URL=`node -p \"require('$CYPRESS_CONFIG').baseUrl\"` && CYPRESS_CONFIG=$CYPRESS_CONFIG start-server-and-test start $BASE_URL cy:run",
-    "test:watch": "yarn run cypress open --config-file cypress/configs/$SHOWCASE_APP.json"
+    "start": "START_COMMAND=`node -p \"require('./cypress/configs/script-config.json').$SCRIPT_ENTRY.startCommand\"` && cd ../$SHOWCASE_APP && $START_COMMAND",
+    "test": "CYPRESS_CONFIG=./cypress/configs/$SHOWCASE_APP.config.ts SCRIPT_ENTRY=`node -p \"'$SHOWCASE_APP'.replaceAll('-', '_');\"` RESPONSE_URL=`node -p \"require('./cypress/configs/script-config.json').$SCRIPT_ENTRY.responseUrl\"` && CYPRESS_CONFIG=$CYPRESS_CONFIG SCRIPT_ENTRY=$SCRIPT_ENTRY start-server-and-test start $RESPONSE_URL cy:run",
+    "test:watch": "yarn run cypress open --config-file cypress/configs/$SHOWCASE_APP.config.ts"
   },
   "devDependencies": {
-    "cypress": "^9.6.1"
+    "cypress": "^10.2.0"
   },
   "dependencies": {
     "cookie-signature": "1.0.6",
