+++ starter-dev-e2e/tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "es5",
     "lib": ["es5", "dom"],
-    "types": ["cypress"],
+    "types": ["cypress", "node"],
     "resolveJsonModule": true
   },
   "include": ["cypress/**/*.ts", "cypress/support/commands.ts"]
